Contour plot changed to a density plot! And it works!
The algorithm is horribly inefficient and confusing to read, but it works very well.
I have found that a smaller number of gradient steps seems to give better results, and
this also makes the algorithm run faster; therefore it may not be worth the effort to
re-write the algorithm. Currently it takes about 5 seconds to make a plot.